Lexer+Diagnostics: Changed how lexing numbers works
Instead of doing it by hand, numbers are now parsed using
long.TryParse() and double.TryParse().

This is done because parsing large or small doubles would
result in extremly inaccurate results.

Another benefit of this is that we can now catch invalid
literals and report a diagnostic for example if they are
too large to fit in the underlying type.